requirements.txt
astroid==2.3.0
Babel==2.7.0
Django==2.2.5
django-phonenumber-field==3.0.1
django-rest-framework==0.1.0
djangorestframework==3.10.3
gunicorn==19.9.0
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
phonenumbers==8.10.19
psycopg2-binary==2.8.3
pylint==2.4.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
